5 задание готово
